# docs/api-contract.openapi.yaml
openapi: 3.1.0
info:
  title: 인테리어 견적 관리 API
  version: 0.2.0
  description: |
    프론트와 백엔드의 단일 진실 공급원. 어떤 코드보다 먼저 수정한다.

    ## 응답 규격
    - **성공**: `{ ok: true, data: T }`
    - **실패**: `{ ok: false, error: string, fieldErrors?: Record<string, string> }`

    ## 인증
    세션 쿠키 기반 (`sid`). `/api/health`, `/api/auth/signup`, `/api/auth/login` 외
    모든 라우트는 로그인 필요. 401 응답 시 클라이언트는 `/login` 으로 리다이렉트.

    ## 변경 절차
    1. 이 문서 먼저 수정
    2. 백엔드 라우트 + repos 구현
    3. 프론트 `api.js` 함수 + 모듈 호출
    4. 양쪽 typecheck 통과 확인

    ## 도메인 노트
    - **견적 채택**: `quotes.status` 와 `line_items.selected_quote_id` 를 의도적으로 분리.
      `POST /api/quotes/{id}/select` 는 line_items.selected_quote_id 만 갱신, status 는 사용자가 수동 관리.
    - **첨부**: 현재 base64 data URL 로 직렬화해 JSON 으로 전송, 5MB 제한.
      추후 `@fastify/multipart` + S3 로 이전 시 별도 라우트 분리.
    - **순서 변경**: phases / spaces / line-items / review-materials 는 별도 reorder 엔드포인트.
      attachments 도 동일.
  contact:
    name: Project repository
    url: https://github.com/your-org/interior-v2
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: http://localhost:3001
    description: 로컬 개발 서버
  - url: https://interior.example.com
    description: 운영 (예시)
...
security:
  - cookieAuth: []
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sid
      description: 세션 쿠키. signup/login 성공 시 발급. health/signup/login 외 모든 라우트는 이 쿠키 필요.
  schemas:
    ApiSuccess:
      type: object
      required: [ok, data]
      properties:
        ok: { type: boolean, enum: [true] }
        data: {}
    ApiError:
      type: object
      required: [ok, error]
      properties:
        ok: { type: boolean, enum: [false] }
        error: { type: string }
        fieldErrors:
          type: object
          additionalProperties: { type: string }
    DeleteSuccess:
      description: 삭제 성공 응답. data.id 는 삭제된 리소스 id.
      allOf:
        - $ref: "#/components/schemas/ApiSuccess"
        - type: object
          required: [data]
          properties:
            data:
              type: object
              required: [id]
              properties:
                id: { type: string }
...
    LineItem:
      type: object
      required: [id, phaseId, label, sortOrder, createdAt, updatedAt]
      properties:
        id: { type: string }
        phaseId: { type: string }
        spaceId: { type: string, nullable: true }
        locationLabel: { type: string, nullable: true }
        workItemLabel: { type: string, nullable: true }
        label: { type: string }
        memo: { type: string, nullable: true }
        selectedQuoteId: { type: string, nullable: true }
        sortOrder: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    Payment:
      type: object
      required: [id, projectId, kind, status, amount, createdAt, updatedAt]
      properties:
        id: { type: string }
        projectId: { type: string }
        contractId: { type: string, nullable: true }
        kind: { type: string, enum: [deposit, progress, final, extra, refund] }
        status: { type: string, enum: [pending, paid, overdue, cancelled] }
        amount: { type: integer }
        dueDate: { type: string, format: date, nullable: true }
        paidAt: { type: string, format: date-time, nullable: true }
        memo: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
  responses:
    Unauthorized:
      description: 로그인 필요
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    NotFound:
      description: 리소스 없음 또는 권한 없음
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    ValidationError:
      description: 입력 검증 실패
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
paths:
...
  /api/phases/{phaseId}/line-items:
    parameters:
      - in: path
        name: phaseId
        required: true
        schema: { type: string }
    post:
      tags: [line-items]
      summary: 견적 항목 생성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [label]
              properties:
                label: { type: string }
                spaceId: { type: string, nullable: true }
                locationLabel: { type: string, nullable: true }
                workItemLabel: { type: string, nullable: true }
                memo: { type: string, nullable: true }
      responses:
        "200":
          description: 생성된 항목
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiSuccess"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/LineItem" }
...
  /api/projects/{projectId}/payments:
    parameters:
      - in: path
        name: projectId
        required: true
        schema: { type: string }
    get:
      tags: [payments]
      summary: 프로젝트 결제 목록
      responses:
        "200":
          description: 결제 목록
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiSuccess"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/Payment" }
    post:
      tags: [payments]
      summary: 결제 생성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kind, amount]
              properties:
                contractId: { type: string, nullable: true }
                kind: { type: string, enum: [deposit, progress, final, extra, refund] }
                status: { type: string, enum: [pending, paid, overdue, cancelled] }
                amount: { type: integer }
                dueDate: { type: string, format: date, nullable: true }
                paidAt: { type: string, format: date-time, nullable: true }
                memo: { type: string, nullable: true }
      responses:
        "200":
          description: 생성된 결제
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiSuccess"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/Payment" }
  /api/payments/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string }
    patch:
      tags: [payments]
      summary: 결제 수정
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                contractId: { type: string, nullable: true }
                kind: { type: string, enum: [deposit, progress, final, extra, refund] }
                status: { type: string, enum: [pending, paid, overdue, cancelled] }
                amount: { type: integer }
                dueDate: { type: string, format: date, nullable: true }
                paidAt: { type: string, format: date-time, nullable: true }
                memo: { type: string, nullable: true }
      responses:
        "200":
          description: 수정된 결제
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiSuccess"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/Payment" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [payments]
      summary: 결제 삭제
      responses:
        "200":
          description: 삭제됨
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeleteSuccess" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/phases/{phaseId}/line-items/reorder:
    parameters:
      - in: path
        name: phaseId
        required: true
        schema: { type: string }
    patch:
      tags: [line-items]
      summary: 견적 항목 순서 변경
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [orderedIds]
              properties:
                orderedIds:
                  type: array
                  items: { type: string }
      responses:
        "200":
          description: 재정렬된 항목 목록
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiSuccess"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/LineItem" }
  /api/line-items/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string }
    patch:
      tags: [line-items]
      summary: 항목 수정
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                phaseId: { type: string }
                label: { type: string }
                spaceId: { type: string, nullable: true }
                locationLabel: { type: string, nullable: true }
                workItemLabel: { type: string, nullable: true }
                memo: { type: string, nullable: true }
                selectedQuoteId: { type: string, nullable: true }
      responses:
        "200":
          description: 수정된 항목
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiSuccess"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/LineItem" }
    delete:
      tags: [line-items]
      summary: 항목 삭제
      responses:
        "200":
          description: 삭제됨
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeleteSuccess" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...